Moved the device module into its own crate
